feat: adding --replace-workflow-config to fully replace workflow configs (from config: directive) with --configfile, instead of merging them (#3381)
<!--Add a description of your PR here-->

Adds a new flag `--replace-workflow-config` to change the behavior of
config file merging: Instead of extended und updating the values in the
config dictionary of the workflow, the config dictionary provided via
CLI replaces the workflow dictionary entirely. Keys that are not
redefined via the CLI dictionary are not present in the final config.

This addresses the request from #730 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a new command‐line option allowing users to fully replace
the workflow configuration with a provided configuration file, ensuring
only the specified settings are applied.
- Enhanced logging now clearly indicates whether configurations are
being extended or replaced, providing clearer insights during execution.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>