fix: 3342 faster touch runs and warning messages for non-existing files (#3398)
### Description

This PR tries to optimize the --touch runs but running the *sleep*
command only for jobs that actually need to touch files.
The *sleep* command is moved into the async *touch* function and run
only if there are files to be created.

Sleeping inside this async should block the entire asyncio event loop,
which is the desired behavior here as the order of files creation is
preserved.

For jobs with non-existing files, a warning messsage is displayed to the
user. A new *log_warning* function is added in the Job class.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a configurable pause during file processing to enhance
operational flexibility.
  
- **Refactor**
- Streamlined the process for checking and processing output files,
ensuring clearer differentiation between file types.
- Improved logging to notify when certain expected files are not
present.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>