feat: files added in rule params with workflow.source_path will be available in used containers  (#3385)
If files are specified under `params` using workflow.source_path, they
are also automatically available in the container. This was achieved by
mounting the nextflow source cache in the container by default. This
achieves a behavior which is rather natural, as the user probably
expects that files which are specified in params using
workflow.source_path are also available in the container.

Due to the fact that the source cache is normally located under
/home/<user>/.cache/snakemake/snakemake/source-cache and the container
thus gets access to the home directory despite the use of the --no-home
flag used in this PR, the documentation explicitly points out that only
trustworthy containers should be used.


### QC

* [ x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

This release improves container workflow execution and documentation,
ensuring a smoother and more secure experience for users.

- **New Features**
- Enhanced container integration now automatically binds local cache
directories to improve workflow performance.
- Added new test scenarios to validate Snakemake execution with
Apptainer deployment.
- **Documentation**
- Updated guidance on file accessibility within container environments,
with added security recommendations to use trusted container images.
- **Tests**
- Introduced additional test functions to cover various Snakemake
execution scenarios.
  - New input and expected output files added for comprehensive testing.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->