fix: edgecases of source deployment in case of remote execution (#3396)
Ensures sources are not uploaded on dry run, and they are cleaned up if
nothing to be done.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Refined deployment processing to ensure that source uploads occur only
during active operations, preventing unintended actions during
simulation or when there are no tasks to process.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->