feat: optionally auto-group jobs via temp files in case of remote execution (#3378)
<!--Add a description of your PR here-->

issue #3355 

creates a new flag to flag files that are only accessible within a group
running on a compute node, and note accessible to the main process
(somewhat reminiscent as how pipes and services are grouped together but
with regular files and directories on paths that are local to that
machin, e.g., that machine's /tmp)

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced support for node-local storage workflows, enabling local
handling of temporary and persistent files.
- Enabled job grouping for tasks that leverage local storage to optimize
performance.

- **Documentation**
- Added a new section detailing job-grouped temporary files to guide
efficient usage of local storage.

- **Refactor**
- Improved error handling and file flag management to prevent conflicts
and ensure smoother operations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>