test: Add pytest markers & skips, update docs to reflect changes (#3417)
<!--Add a description of your PR here-->
- Some tests that that are related to s3 storage keep failing locally. I
believe these require the S3 credentials (available as secretes in CI)
so I've marked them with pytest as `needs_s3` and added a section to the
docs regarding this.
- I could instead have a marker skip automatically but unsure how this
works at the moment and dont want to have it start skipping tests in
CI/CD.

- I've also updated the `.test_durations` files since it hasnt been
updated in some time, and there are more tests.

Side note: the docs havent updated due to the src layout not registering
in ReadtheDocs. fixed: #3419

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Documentation**
- Introduced a new guide on using test markers, providing clear
instructions for running or excluding tests based on environment needs.

- **Tests**
- Revised and reorganized test configurations with adjusted timings and
added numerous new test cases.
- Enhanced test annotations to indicate dependencies on external
resources and platform-specific conditions for improved clarity during
execution.
- Added support for macOS in the testing workflow, expanding the
environment options.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->