feat: Fix keep_local in storage directive and more freedom over remote retrieval behaviour (#3410)
This PR enhances the behaviour of `keep_local` and `retrieve` when using
remote storage plugins:
* `keep_local` was not always respected when used in a storage directive
(see https://github.com/snakemake/snakemake/issues/2844 e.g.)
* This allows `keep_local` to be specified in the storage directive
(which overrides the CLI argument) and in the call to `storage` (which
overrides the specification, if any, in the storage directive)
* The default behaviour for retrieving remote files is to always do so
unless `retrieve=False` is used in the call to `storage`
* This PR moves the behaviour more inline with `keep_local` i.e. a
default behaviour can be specified with a new CLI argument
`--not-retrieve-local` (not a very graceful name, open to suggestions!)
* The default behaviour is to keep `retrieve=True` as is currently the
case in snakemake but it can be overridden in a storage directive
* This requires some small changes in
`snakemake-interface-storage-plugins` which I will open another PR for

This PR also changes the behaviour when defining a tagged storage
provider without specifying the default:
* If a tagged storage provider is defined but not the default one, then
the settings of the tagged provider are not properly respected as the
underlying provider is registered later only when the `storage_object`
is accessed i.e. it has no access to the tagged settings
* This is changed by checking, when a tagged storage provider is
registered, if the underlying provider has not yet been registered then
it is registered using the same settings as the tagged provider

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Expanded documentation to guide users on managing remote files and
controlling local file retention through new options.

- **New Features**
  - Introduced a command-line option to disable remote file downloads.
- Enhanced the workflow's cleanup process to preserve local files when
desired.
- Added configurable settings for controlling file retrieval and local
storage behavior.

- **Tests**
- Included new example workflows and automated tests to showcase and
verify the updated storage and retrieval functionalities.
- Added tests for local storage retention and retrieval behaviors under
various conditions.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>