docs: update broken links documentation (#3437)
<!--Add a description of your PR here-->

The src directory change broke several links in the documentation, this
PR corrects them.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Updated component references in the documentation to align with the
new directory organization, ensuring all links are accurate and
consistent.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->