feat: Add option to print redacted file names (#3089)
Addresses issue https://github.com/snakemake/snakemake/issues/3087.

Adds some functionality so that if a file is a storage object, it can
print a censored version of the query (`print_query`) rather than just
the query itself incase it contains sensitive information in the query
that should not be printed.

I suppose it would make sense to at least put something in the
documentation (in `snakemake-interface-storage-plugins` maybe?) to
explain that the `safe_print` method can be used (as an aside, I tested
this with the `_safe_to_print_url` method of
`snakemake-storage-plugin-xrootd` but figured that method name is
probably not general enough so I can rename it for that plugin).

Note this also slightly changes the message when printing an input file
as previously it would always print `(retrieve from storage)` even if
`retrieve=False`.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enabled configurable integration for xrootd-based storage, enhancing
file handling workflows.
- **Refactor**
- Improved messages and output clarity for storage operations, offering
more precise file processing instructions.
- **Tests**
- Added validations to confirm that sensitive credentials remain
concealed.
- **Chores**
- Updated continuous integration triggers to restrict execution based on
repository name.
	- Added a new dependency for enhanced xrootd storage capabilities.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>