Docs/migration (#3446)
### Description

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Introduced a new migration guide outlining updates for version 9,
including changes in specifying custom log handlers via new designated
options. The update is designed to have minimal impact on most users.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->