fix: group job error call (#3448)
This wasn't caught earlier, likely due to how we check for tests that
expect failure. The `test_group_jobs_fail` test was failing, but it was
b/c of this issue. Perhaps should check for the expected exception.
### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced error logging to display grouped job errors in a sorted,
consistent order for clearer troubleshooting.
- Introduced tailored error logging during job scheduling, ensuring that
errors for grouped jobs are handled separately for improved clarity.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>