docs: fix profiles linking, add Sphinx rst syntax infos (#3453)
Since the inclusion of the migration guide, the linking to the
`profiles` section on the CLI page was broken. This guide contains a
heading of the name `profiles`, and Sphinx auto-generates links from
headings. So this PR moves to the hyperlink target name
`executing-profiles`, as this is what the documentation guidelines
recommend, anyways.

In addition, it adds some essential info on the format used for the
docs, with linkouts to the Sphinx documentation.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Updated label references from "profiles" to "executing-profiles"
across several documentation sections for improved clarity on execution
configurations.
- Revised reference links in migration and configuration guides to
direct users to the updated execution profiles information.
- Enhanced contributor guidelines with additional details and links on
using Sphinx and reStructuredText for improved documentation clarity.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->