fix: Resolve issues with hashlib.md5() on systems in FIPS mode (#3522)
<!--Add a description of your PR here-->

This change adds `usedforsecurity=False` to calls of `hashlib.md5()` to
resolve issues with the availability of the MD5 hash algorithm on
systems with (United States) Federal Information Processing Standards
(FIPS) mode enabled. Snakemake does not work on systems with this
security standard because the MD5 algorithm is considered unsafe. For
more information, please see #3504.

Resolves #3504

### QC
<!-- Make sure that you can tick the boxes below. -->

* [X] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [X] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Refined several internal hashing routines to clearly indicate that
they are not used for cryptographic security, ensuring more consistent
behavior across the system.
- **Tests**
- Updated test validations to align with the revised internal
configuration.

No changes have been made to the public interfaces, so you should
experience uninterrupted functionality.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->