docs: add argument for `--dag` flag (#3524)
It appears that the CLI `--dag` now takes arguments, which is omitted in
the documentation. I've added it in the example.

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Updated the command instructions for visualizing the workflow graph
with Snakemake, ensuring users reference the modified directive when
generating the DAG visualization.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->