fix: log scheduler waiting to debug, make statements more accurate (#3547)
<!--Add a description of your PR here-->
Fixes #3471.

Provide more accurate log info when scheduler is waiting for
jobs/resources. Move these statements to debug to prevent flooding of
stderr.
### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Refactor**
- Improved scheduler logging messages to more accurately indicate
whether it is waiting for running jobs to complete or for more
resources, and changed these messages to debug-level logs.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->