docs: rewrite the mpi job execution section (#3538)
This pull request tries to make the explanations in this section a bit
clearer and provides a number of useful linkouts, especially to the Open
MPI docs, but also to the snakemake executor plugins. Along the way, it
also fixes the syntax of existing links.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Improved and expanded the MPI support documentation for Snakemake
workflows, including clearer explanations, updated examples, and
guidance on customizing MPI launcher commands.
- Added references to external MPI and cluster documentation for further
assistance.
- Enhanced instructions for overriding MPI resources and passing
additional parameters.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->